+++ Job_Portail_Spring_Final/src/main/java/com/example/security/Other/Job/JobRepository.java
@@ -1,11 +1,7 @@
 package com.example.security.Other.Job;
 
-// import com.example.security.Other.Job.Job;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-// package com.example.security.Other.JobSkill;
-
 @Repository
-public interface JobRepository extends JpaRepository<Job, Long> {}
-
+public interface JobRepository extends JpaRepository<Job, Integer> {}
