+++ Job_Portail_Spring_Final/src/main/java/com/example/security/Other/JobSkill/JobSkillDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
     @NoArgsConstructor
     @AllArgsConstructor
     public class JobSkillDto {
-        private Long skillId;
+        private Integer skillId;
         private String skillName;
         private Boolean required;
     }
