+++ Job_Portail_Spring_Final/src/main/java/com/example/security/Other/JobSkillId/JobSkillId.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @Builder
 
 public class JobSkillId implements Serializable {
-    private Long job;
-    private Long skill;
+    private Integer job;
+    private Integer skill;
 }
