+++ Job_Portail_Spring_Final/src/main/java/com/example/security/Other/Skill/SkillRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SkillRepository extends JpaRepository<Skill, Long> {
+public interface SkillRepository extends JpaRepository<Skill, Integer> {
     // Optional: find by skill name, e.g.:
     // Optional<Skill> findByNameIgnoreCase(String name);
 }
